+++ database/migrations/2021_11_28_015748_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
 
             $table->string('order_number');
             $table->enum('status', ['pending','processing','completed','decline'])->default('pending');
-            $table->float('grand_total');
+            $table->unsignedBigInteger('grand_total');
             $table->integer('item_count');
             $table->boolean('is_paid')->default(false);
             $table->enum('payment_method', ['cash_on_delivery', 'OM','MOMO'])->default('cash_on_delivery');
